# CLN: Use read_file from geodataframe in geoseries