--- a/geopandas/geoseries.py
+++ b/geopandas/geoseries.py
@@ -78,7 +78,6 @@ class GeoSeries(GeoPandasBase, Series):
 
         Parameters
         ----------
-
         filename : str
             File path or file handle to read from. Depending on which kwargs
             are included, the content of filename may vary. See
@@ -88,15 +87,11 @@ class GeoSeries(GeoPandasBase, Series):
             access multi-layer data, data stored within archives (zip files),
             etc.
         """
-        import fiona
-        geoms = []
-        with fiona.open(filename, **kwargs) as f:
-            crs = f.crs
-            for rec in f:
-                geoms.append(shape(rec['geometry']))
-        g = GeoSeries(geoms)
-        g.crs = crs
-        return g
+
+        from geopandas import GeoDataFrame
+        df = GeoDataFrame.from_file(filename, **kwargs)
+
+        return GeoSeries(df.geometry, crs=df.crs)
 
     @property
     def __geo_interface__(self):
@@ -113,8 +108,8 @@ class GeoSeries(GeoPandasBase, Series):
     def to_file(self, filename, driver="ESRI Shapefile", **kwargs):
         from geopandas import GeoDataFrame
         data = GeoDataFrame({"geometry": self,
-                          "id":self.index.values},
-                          index=self.index)
+                             "id": self.index.values},
+                            index=self.index)
         data.crs = self.crs
         data.to_file(filename, driver, **kwargs)
 
@@ -335,5 +330,6 @@ class GeoSeries(GeoPandasBase, Series):
     def __sub__(self, other):
         """Implement - operator as for builtin set type"""
         return self.difference(other)
+
 
 GeoSeries._create_indexer('cx', _CoordinateIndexer)