@jorisvandenbossche 2019-01-02
I fully agree we should try to minimize this duplicated code. But was thinking: one advantage of the current code compared to your PR is that it *only* reads the geoms, and doesn't parse all the properties. 
I personally don't care much about that use case, but I suppose that this is the reason this uses a custom loop instead of reusing `read_file`.