@jorisvandenbossche 2019-01-07
Hmm, that's indeed not much difference. 
It might be that currently parsing the geometries takes almost all time compared to parsing the attributes.

So I am fine with this PR.

Although that said, I am personally not really a fan of the `from_` methods, since they duplicate `geopandas.read_file`, so we could maybe also consider deprecating them (but that's a separate issue)