@jdmcbr 2019-01-07
> Although that said, I am personally not really a fan of the from_ methods, since they duplicate geopandas.read_file, so we could maybe also consider deprecating them (but that's a separate issue)

I agree on this. At least, I've never seen much value in them.

Timing with different file types doesn't change the story (I dropped the number of points in the file a bit just because it was a little slow to run all three file types with so many points when I was playing around).

master:
```

[100.00%] ··· io.Bench.time_read_series_from_frame                            ok
[100.00%] ··· ======= ==========
                ext             
              ------- ----------
                .shp   471±7ms  
               .json   654±20ms 
               .gpkg   474±8ms  
              ======= ==========
```

this branch:
```
[100.00%] ··· io.Bench.time_read_series_from_frame                            ok
[100.00%] ··· ======= ==========
                ext             
              ------- ----------
                .shp   496±20ms 
               .json   662±10ms 
               .gpkg   476±20ms 
              ======= ==========
```