@martinfleis 2019-05-19
I am going to merge this. There is a general agreement that this PR is useful. Concerns are only about possible deprecation of `from_` methods (which I support), but that is indeed another discussion.